Rename `rflx/parser` to `rflx/specification`
Ref. #252
